add some minor edits

+++ javaAdvanced-21/src/main/java/hw/dao/BasketRepo.java
@@ -3,6 +3,6 @@ package hw.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import hw.domain.Basket;
 
-public interface BasketRepo extends JpaRepository<Basket, Long>{
+public interface BasketRepo extends JpaRepository<Basket, Integer>{
 
 }
